Fix jest test app test
